requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
langchain
langchain-groq
langchain-core
langchain-chroma
langchain-ollama
lazypredict
python-dotenv
sqlite3